moving the url-modifiying test out of BaseDataSourceTest (#1039)
